backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
insightface==0.7.3
opencv-python==4.11.0.86
numpy==1.26.4
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
django-cors-headers==4.3.1
onnxruntime==1.16.3
faiss-cpu==1.7.4
openpyxl==3.1.2
reportlab==4.0.7
django-filter==23.3
daphne==4.0.0
python-dotenv==1.0.0
scipy==1.11.4
requests==2.31.0
celery==5.3.4
django-celery-beat==2.5.0
